+++ __browser_tests__/fixtures/relay-app/__generated__/UpdateUserMutation.graphql.js
@@ -1,0 +1,127 @@
+/**
+ * @flow
+ * @relayHash 932e73263fced7b11cdee21a62483c64
+ */
+
+/* eslint-disable */
+
+'use strict';
+
+/*::
+import type { ConcreteRequest } from 'relay-runtime';
+export type UpdateUserInput = {
+  id: string,
+  name?: ?string,
+  clientMutationId?: ?string,
+};
+export type UpdateUserMutationVariables = {|
+  input: UpdateUserInput
+|};
+export type UpdateUserMutationResponse = {|
+  +updateUser: ?{|
+    +user: ?{|
+      +id: string,
+      +name: ?string,
+    |}
+  |}
+|};
+export type UpdateUserMutation = {|
+  variables: UpdateUserMutationVariables,
+  response: UpdateUserMutationResponse,
+|};
+*/
+
+
+/*
+mutation UpdateUserMutation(
+  $input: UpdateUserInput!
+) {
+  updateUser(input: $input) {
+    user {
+      id
+      name
+    }
+  }
+}
+*/
+
+const node/*: ConcreteRequest*/ = (function(){
+var v0 = [
+  {
+    "kind": "LocalArgument",
+    "name": "input",
+    "type": "UpdateUserInput!",
+    "defaultValue": null
+  }
+],
+v1 = [
+  {
+    "kind": "LinkedField",
+    "alias": null,
+    "name": "updateUser",
+    "storageKey": null,
+    "args": [
+      {
+        "kind": "Variable",
+        "name": "input",
+        "variableName": "input",
+        "type": "UpdateUserInput!"
+      }
+    ],
+    "concreteType": "UpdateUserPayload",
+    "plural": false,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "user",
+        "storageKey": null,
+        "args": null,
+        "concreteType": "User",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "id",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "name",
+            "args": null,
+            "storageKey": null
+          }
+        ]
+      }
+    ]
+  }
+];
+return {
+  "kind": "Request",
+  "operationKind": "mutation",
+  "name": "UpdateUserMutation",
+  "id": null,
+  "text": "mutation UpdateUserMutation(\n  $input: UpdateUserInput!\n) {\n  updateUser(input: $input) {\n    user {\n      id\n      name\n    }\n  }\n}\n",
+  "metadata": {},
+  "fragment": {
+    "kind": "Fragment",
+    "name": "UpdateUserMutation",
+    "type": "Mutation",
+    "metadata": null,
+    "argumentDefinitions": v0,
+    "selections": v1
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "UpdateUserMutation",
+    "argumentDefinitions": v0,
+    "selections": v1
+  }
+};
+})();
+// prettier-ignore
+(node/*: any*/).hash = 'a13c16598355408b91ea8b166f5f6756';
+module.exports = node;