+++ graphql-relay-subscription.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/server/graphql-relay-subscription.js')