+++ src/server/__tests__/__fixtures__/schema/create-user-query.js
@@ -1,0 +1,7 @@
+const UserType = require('./user-type')
+
+module.exports = (userResolve) => ({
+  type: UserType,
+  args: {},
+  resolve: userResolve
+})