+++ test-browser/fixtures/babel-relay-plugin.js
@@ -1,3 +1,0 @@
-const schema = require('../../test/fixtures/graphql-schema.json')
-
-module.exports = require('babel-relay-plugin')(schema)