+++ test-browser/fixtures/debug-server.js
@@ -1,6 +1,0 @@
-require('./karma-primus.js')
-
-require('http').createServer().listen(8181, function (err) {
-  if (err) throw err
-  console.log('listening on 8181')
-})