+++ test-browser/fixtures/root-container.jsx
@@ -1,9 +1,0 @@
-var React = require('react')
-var Relay = require('react-relay')
-var UserContainer = require('./user-container.jsx')
-var UserRoute = require('./user-route.jsx')
-
-module.exports = <Relay.RootContainer
-  Component={UserContainer}
-  route={new UserRoute()}
-/>